misc/HtmlTestSuite/results/test-1366.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Shadow DOM Test: A_05_01_03</title>
<link href="mailto:sgrekhov@unipro.ru" rel="author" title="Sergey G. Grekhov"/>
<link href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#event-retargeting" rel="help"/>
<meta content="Event Retargeting:Event retargeting for fallback content" name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../../../../html/resources/common.js"/>
<script src="../../../resources/shadow-dom-utils.js"/>
</head>
<body>
<div id="log"/>
<script>
var A_05_01_03_T01 = async_test('A_05_01_03_T01');

A_05_01_03_T01.step(unit(function (ctx) {
    var d = newRenderedHTMLDocument(ctx);

    d.body.innerHTML = '' +
    '&lt;div id="main"&gt;' +
        '&lt;div id="shadow-root"&gt;' +
            '&lt;span&gt;1&lt;/span&gt;' +
            '&lt;span&gt;2&lt;/span&gt;' +
            '&lt;span&gt;3&lt;/span&gt;' +
        '&lt;/div&gt;' +
    '&lt;/div&gt;';

    var ul = d.querySelector('#shadow-root');
    var s = ul.attachShadow({mode: 'open'});

    //make shadow subtree
    var div = document.createElement('div');
    div.innerHTML = '&lt;slot name="shadow"&gt;&lt;span id="flbk"&gt;Fallback item&lt;/span&gt;&lt;/slot&gt;';
    s.appendChild(div);

    d.body.addEventListener('click', A_05_01_03_T01.step_func(function (event) {
        assert_equals(event.target.getAttribute('id'), 'shadow-root', 'Information about ' +
                'event target crossing the shadow boundaries should be adjusted for the fallback ' +
                'content');
    }), false);

    var event = d.createEvent('HTMLEvents');
    event.initEvent ("click", true, false);
    s.querySelector('#flbk').dispatchEvent(event);

    A_05_01_03_T01.done();
}));
</script>
</body>
</html>
